Selection Sort project
This file contains all the requirements of the project. My random number
generator file is in RandomNumberGenerator. My Selection sort sorts both
the number in my document and arrays. This is seen because my reader
file reads the numbers within my document and then writes them to arrays
which my SelectionSort.java sorts.